requirements.txt
alembic==1.0.11
Click==7.0
dominate==2.3.5
Flask==1.1.1
Flask-Bootstrap4==4.0.2
Flask-FlatPages==0.7.1
Flask-Login==0.4.1
Flask-Markdown==0.3
Flask-Migrate==2.5.2
flask-shell==0.1.3
Flask-SQLAlchemy==2.4.0
Frozen-Flask==0.15
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
migrate==0.3.8
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
PyYAML==5.1.2
six==1.12.0
SQLAlchemy==1.3.6
visitor==0.1.3
Werkzeug==0.15.5
